+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to the versioning scheme outlined in the [README.md](RE
 
 ## [2.05.0.2.0]
 
+- Updates to the logging of transaction events (#3139).
+
 ### IMPORTANT! READ THIS FIRST
 
 Please read the following **WARNINGs** in their entirety before upgrading.